# backend/src/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "SmartMarketOOPS API",
    "description": "API documentation for SmartMarketOOPS Trading Platform",
    "version": "1.0.0",
    "contact": {
      "name": "SmartMarketOOPS Team",
      "email": "109647440+abhaskumarrr@users.noreply.github.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3006",
      "description": "Development Server"
    },
    {
      "url": "https://api.smartmarketoops.com",
      "description": "Production Server"
    }
  ],
...
  "paths": {
...
    "/api/delta-trading/bots": {
      "get": {
        "tags": ["delta-trading", "bots"],
        "summary": "Get all trading bots",
        "description": "Get status of all trading bots",
        "operationId": "getAllBots",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of all trading bots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotsListResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to get bots",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": ["delta-trading", "bots"],
        "summary": "Create a new trading bot",
        "description": "Create a new trading bot with the specified configuration",
        "operationId": "createBot",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "description": "Trading bot configuration",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BotConfig"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Trading bot created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/BotCreatedResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid bot configuration",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "500": {
            "description": "Failed to create bot",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "BotsListResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/BotStatus"
            }
          },
          "meta": {
            "type": "object",
            "properties": {
              "total": {
                "type": "number",
                "example": 5
              },
              "running": {
                "type": "number",
                "example": 3
              },
              "stopped": {
                "type": "number",
                "example": 1
              },
              "paused": {
                "type": "number",
                "example": 0
              },
              "error": {
                "type": "number",
                "example": 1
              }
            }
          },
          "timestamp": {
            "type": "number",
            "example": 1622547868000
          }
        }
      },
      "BotStatus": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "bot-123"
          },
          "name": {
            "type": "string",
            "example": "BTC Momentum Bot"
          },
          "status": {
            "type": "string",
            "enum": ["running", "stopped", "paused", "error"],
            "example": "running"
          },
          "symbol": {
            "type": "string",
            "example": "BTCUSDT"
          },
          "strategy": {
            "type": "string",
            "example": "momentum"
          },
          "capital": {
            "type": "number",
            "example": 1000
          },
          "activePositions": {
            "type": "number",
            "example": 1
          },
          "profitLoss": {
            "type": "number",
            "example": 50.25
          },
          "runningTime": {
            "type": "number",
            "example": 3600
          },
          "lastUpdate": {
            "type": "number",
            "example": 1622547868000
          }
        }
      },
...
      "BotConfig": {
        "type": "object",
        "required": ["id", "name", "symbol"],
        "properties": {
          "id": {
            "type": "string",
            "example": "bot-123"
          },
          "name": {
            "type": "string",
            "example": "BTC Momentum Bot"
          },
          "symbol": {
            "type": "string",
            "example": "BTCUSDT"
          },
          "strategy": {
            "type": "string",
            "example": "momentum"
          },
          "capital": {
            "type": "number",
            "example": 1000
          },
          "leverage": {
            "type": "number",
            "example": 3
          },
          "riskPerTrade": {
            "type": "number",
            "example": 2
          },
          "maxPositions": {
            "type": "number",
            "example": 3
          },
          "stopLoss": {
            "type": "number",
            "example": 5
          },
          "takeProfit": {
            "type": "number",
            "example": 10
          },
          "enabled": {
            "type": "boolean",
            "example": true
          }
        }
      },
      "BotCreatedResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "data": {
            "type": "object",
            "properties": {
              "botId": {
                "type": "string",
                "example": "bot-123"
              },
              "config": {
                "$ref": "#/components/schemas/BotConfig"
              }
            }
          },
          "message": {
            "type": "string",
            "example": "Trading bot created successfully"
          },
          "timestamp": {
            "type": "number",
            "example": 1622547868000
          }
        }
      },
      "ErrorResponse": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": false
          },
          "error": {
            "type": "string",
            "example": "Failed to create bot"
          },
          "message": {
            "type": "string",
            "example": "Invalid configuration: missing required field 'symbol'"
          }
        }
      }
    }
  }
} 